+++ src/main/java/com/crud/repository/UserLoginRepository.java
@@ -4,8 +4,9 @@ import com.crud.entity.UserLogin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface UserLoginRepository extends JpaRepository<UserLogin, Integer> {
-    Optional<UserLogin> findUserByDi(BigDecimal di);
+    Optional<UserLogin> findUserByDi(BigInteger di);
 }
