+++ src/main/java/com/crud/services/UserLoginService.java
@@ -4,10 +4,11 @@ import com.crud.entity.UserLogin;
 import org.springframework.http.ResponseEntity;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface UserLoginService {
-    ResponseEntity<Object> findUserByDi(BigDecimal di);
+    ResponseEntity<Object> findUserByDi(BigInteger di);
     ResponseEntity<Object> save (UserLogin user);
     ResponseEntity<Object> updateUser (UserLogin user);
-    ResponseEntity<Object> delete(BigDecimal di);
+    ResponseEntity<Object> delete(BigInteger di);
 }
